Created a new project for the level editor (Afterglow Editor). Implemented an ImGui dockspace with some Renderer stats (draw calls count for now) and a viewport that's resizable and processes inputs if it's focused and hovered. Updated the framebuffer to handle resizing and did maintenance work related to updating some data types for few sources.

+++ Afterglow/src/Afterglow/Core/Events/WindowEvents.h
@@ -10,7 +10,7 @@ namespace Afterglow
 	class WindowResizeEvent : public Event
 	{
 	public:
-		WindowResizeEvent(uint32_t width, uint32_t height)
+		WindowResizeEvent(uint16_t width, uint16_t height)
 			: m_Width(width), m_Height(height) 
 		{
 		}
@@ -28,11 +28,11 @@ namespace Afterglow
 			return ss.str();
 		}
 
-		uint32_t GetWidth() const { return m_Width; }
-		uint32_t GetHeight() const { return m_Height; }
+		uint16_t GetWidth() const { return m_Width; }
+		uint16_t GetHeight() const { return m_Height; }
 	
 	private:
-		uint32_t m_Width, m_Height;
+		uint16_t m_Width, m_Height;
 	};
