+++ Afterglow/src/Afterglow/Platform/Windows/Window/WindowsWindow.cpp
@@ -87,7 +87,7 @@ namespace Afterglow
 				data.Width = width;
 				data.Height = height;
 
-				WindowResizeEvent WREvent((uint32_t)width, (uint32_t)height);
+				WindowResizeEvent WREvent((uint16_t)width, (uint16_t)height);
 				data.EventCallback(WREvent);
 			}
 		);
